Enhancement: Extract IsReadableTest
